some tweaks to install script and yazi

+++ fullinstall.sh
@@ -94,6 +94,7 @@ APPS=(
   "wlogout"
   "yazi"
   "zoxide"
+  "zsh"
 )
 # Install apps only if not already installed
 for app in "${APPS[@]}"; do
@@ -110,18 +111,39 @@ for app in "${APPS[@]}"; do
 done
 echo "All requested apps processed."
 
-wal -i ~/wallpaper/Sword Black.png -n
+wal -i ~/wallpaper/assassin\'sCreed.png -n
 systemctl enable bluetooth
 systemctl --user enable pipewire.service pipewire-pulse.service
 systemctl --user start pipewire.service pipewire-pulse.service
 systemctl --user enable brightnessctl
 systemctl --user start brightnessctl
-sudo cp -r -f ~/dotfiles/.config/* ~/.config/
-sudo cp -r -f ~/dotfiles/.zshrc ~/
-sudo cp -r -f ~/dotfiles/nixwd-home ~/nixwd-home
+# Manages dotfiles
+if command -v stow &> /dev/null; then
+  read -p "Do you want to add the dotfiles to their respective location using stow or do you want to hard copy them? (y/n): " -n 1 -r
+  echo  # Add a newline after single-character input
+  if [[ $REPLY =~ ^[Yy]$ ]]; then
+    cd ~/dotfiles
+    stow .
+    cd -
+  else
+    sudo cp -r -f ~/dotfiles/.config/* ~/.config/
+    sudo cp -r -f ~/dotfiles/.zshrc ~/
+    sudo cp -r -f ~/dotfiles/nixwd-home ~/
+  fi
+fi
+# INSTALLING nix
 if ! command -v nix &> /dev/null; then
   echo "[INSTALLING] NIX..."
   sh <(curl -L https://nixos.org/nix/install) --daemon
+fi
+if command -v zsh &> /dev/null; then
+  sudo chsh -s /usr/bin/zsh $USER
+fi
+if command -v yazi &> /dev/null; then
+  echo "Configure yazi, for theme install: "
+  echo "ya pack -a yazi-rs/flavors:catppuccin-frappe"
+  echo "For mounting disks install: "
+  echo "ya pack -a yazi-rs/plugins:mount"
 fi
 tldr -u
 notify-send "Open Terminal with MOD+return" "Hello $USER,\nWelcome to your new Arch install\n-EF"
